+++ cloudpathlib/cloudpath.py
@@ -717,6 +717,19 @@ class CloudPath(metaclass=CloudPathMeta):
             f"overwrite."
         )
 
+    # ===========  pydantic integration special methods ===============
+    @classmethod
+    def __get_validators__(cls):
+        """Pydantic special method. See
+        https://pydantic-docs.helpmanual.io/usage/types/#custom-data-types"""
+        yield cls._validate
+
+    @classmethod
+    def _validate(cls, value: Any):
+        """Used as a Pydantic validator. See
+        https://pydantic-docs.helpmanual.io/usage/types/#custom-data-types"""
+        return cls(value)
+
 
 # The function resolve is not available on Pure paths because it removes relative
 # paths and symlinks. We _just_ want the relative path resolution for